Refactor: 예외 index.ts 적용
상동
