+++ server/src/domain/auth/service/auth.service.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Injectable, UnauthorizedException } from '@nestjs/common';
+import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JwtService } from '@nestjs/jwt';
 import { Bcrypt } from 'src/utils/bcrypt';
@@ -8,7 +8,7 @@ import { LoginRequestDto } from '../dto/login-request.dto';
 import { UserResponseDto } from '../dto/user-response.dto';
 import { JoinRequestDto } from '../dto/join-request.dto';
 import { UserBuilder } from '../../../builder';
-import { UserException } from '../../../exception/user.exception';
+import { UserException } from '../../../exception';
 
 @Injectable()
 export class AuthService {
@@ -29,14 +29,12 @@ export class AuthService {
 
   async join(joinRequestDto: JoinRequestDto) {
     const { nickname, email, password } = joinRequestDto;
-    const baseImageURL = '디폴트 이미지 주소 자리';
     const isExistUser = await this.authRepository.exists(joinRequestDto);
     if (isExistUser) throw UserException.userIsExist();
     const user: User = new UserBuilder()
       .setNickName(nickname)
       .setEmail(email)
       .setPassword(Bcrypt.hash(password))
-      .setImageURL(baseImageURL)
       .build();
     await this.authRepository.save(user);
     return true;
@@ -44,7 +42,7 @@ export class AuthService {
 
   async getUserInfo(email: string) {
     const user: User = await this.authRepository.findUserByEmail(email);
-    if (!user) throw new UnauthorizedException();
+    if (!user) throw UserException.userNotFound();
     return new UserResponseDto(user);
   }
 }
